+++ gcc/rust/Make-lang.in
@@ -225,6 +225,7 @@ GRS_OBJS = \
     rust/rust-compile-struct-field-expr.o \
     rust/rust-constexpr.o \
     rust/rust-compile-base.o \
+    rust/rust-compile-drop.o \
     rust/rust-tree.o \
     rust/rust-compile-context.o \
     rust/rust-export-metadata.o \
