+++ gcc/rust/backend/rust-compile-pattern.cc
@@ -17,45 +17,24 @@
 // <http://www.gnu.org/licenses/>.
 
 #include "rust-compile-pattern.h"
+#include "print-tree.h"
+#include "rust-compile-drop.h"
 #include "rust-compile-expr.h"
 #include "rust-compile-resolve-path.h"
-#include "rust-constexpr.h"
 #include "rust-compile-type.h"
-#include "print-tree.h"
+#include "rust-constexpr.h"
 #include "rust-diagnostics.h"
 #include "rust-hir-pattern-abstract.h"
 #include "rust-hir-pattern.h"
+#include "rust-hir-trait-reference.h"
+#include "rust-hir-type-bounds.h"
+#include "rust-lang-item.h"
 #include "rust-system.h"
 #include "rust-tyty.h"
-#include "rust-hir-type-bounds.h"
-#include "rust-hir-trait-reference.h"
-#include "rust-lang-item.h"
 #include "tree.h"
 
 namespace Rust {
 namespace Compile {
-
-static bool
-type_has_drop_impl (Context *ctx, TyTy::BaseType *ty)
-{
-  auto drop_lang_item
-    = ctx->get_mappings ().lookup_lang_item (LangItem::Kind::DROP);
-
-  if (!drop_lang_item.has_value ())
-    return false;
-
-  DefId drop_id = drop_lang_item.value ();
-
-  auto candidates = Resolver::TypeBoundsProbe::Probe (ty);
-  for (auto &candidate : candidates)
-    {
-      Resolver::TraitReference *trait_ref = candidate.first;
-      if (trait_ref != nullptr && trait_ref->get_defid () == drop_id)
-	return true;
-    }
-
-  return false;
-}
 
 void
 CompilePatternCheckExpr::visit (HIR::PathInExpression &pattern)
@@ -1377,7 +1356,7 @@ CompilePatternLet::visit (HIR::IdentifierPattern &pattern)
       drop_ty = ref_ty->get_base ();
     }
 
-  if (!type_has_drop_impl (ctx, drop_ty))
+  if (!CompileDrop::type_has_drop_impl (ctx, drop_ty))
     return;
 
   if (!pattern.has_subpattern () && !pattern.get_is_ref ())
